Stats19 checkbox on /around page.
This should ideally be done with proper backend queries,
but for a demo, we are instead overriding the "Show old"
functionality, as this is functionally equivalent for the
current dataset.

Longer term, we would do something similar, but with a
cobrand_filter parameter, handled in similar places to
the all_pins param (e.g. ::AC::Around, ::Map, ::DBRS::Problem)