‘require’ Script module, not ‘use’.
By using ‘use’, the package was loaded at compile time when the Smidsy
cobrand package was registered, which meant it tried to load a cobrand
before all the cobrands were loaded, leading not to an error but a new
empty cobrand package being created. This causes confusion when it was
then used in place of the actual cobrand package (because it was lower
case, it appeared after the actual package in Module::Pluggable’s list
as long as the version of M::P was recent enough).